requirements.txt
fastapi
uvicorn
psycopg2-binary
python-dotenv
jinja2
pandas
